hotrestart: extract SharedMemory and ProcessSharedMutex and use bazel to select between real and no-op HotRestartImpl code
