+++ source/exe/hot_restart.cc
@@ -3,9 +3,7 @@
 #include <signal.h>
 #include <sys/mman.h>
 
-#if defined(__linux__)
 #include <sys/prctl.h>
-#endif
 
 #include <sys/socket.h>
 #include <sys/types.h>
@@ -28,74 +26,6 @@
 namespace Envoy {
 namespace Server {
 
-// Increment this whenever there is a shared memory / RPC change that will prevent a hot restart
-// from working. Operations code can then cope with this and do a full restart.
-const uint64_t SharedMemory::VERSION = 8;
-
-SharedMemory& SharedMemory::initialize(Options& options) {
-  int flags = O_RDWR;
-  std::string shmem_name = fmt::format("/envoy_shared_memory_{}", options.baseId());
-  if (options.restartEpoch() == 0) {
-    flags |= O_CREAT | O_EXCL;
-
-    // If we are meant to be first, attempt to unlink a previous shared memory instance. If this
-    // is a clean restart this should then allow the shm_open() call below to succeed.
-    shm_unlink(shmem_name.c_str());
-  }
-
-  int shmem_fd = shm_open(shmem_name.c_str(), flags, S_IRUSR | S_IWUSR);
-  if (shmem_fd == -1) {
-    PANIC(fmt::format("cannot open shared memory region {} check user permissions", shmem_name));
-  }
-
-  if (options.restartEpoch() == 0) {
-    int rc = ftruncate(shmem_fd, sizeof(SharedMemory));
-    RELEASE_ASSERT(rc != -1);
-    UNREFERENCED_PARAMETER(rc);
-  }
-
-  SharedMemory* shmem = reinterpret_cast<SharedMemory*>(
-      mmap(nullptr, sizeof(SharedMemory), PROT_READ | PROT_WRITE, MAP_SHARED, shmem_fd, 0));
-  RELEASE_ASSERT(shmem != MAP_FAILED);
-
-  if (options.restartEpoch() == 0) {
-    shmem->size_ = sizeof(SharedMemory);
-    shmem->version_ = VERSION;
-    shmem->initializeMutex(shmem->log_lock_);
-    shmem->initializeMutex(shmem->access_log_lock_);
-    shmem->initializeMutex(shmem->stat_lock_);
-    shmem->initializeMutex(shmem->init_lock_);
-  } else {
-    RELEASE_ASSERT(shmem->size_ == sizeof(SharedMemory));
-    RELEASE_ASSERT(shmem->version_ == VERSION);
-  }
-
-  // Here we catch the case where a new Envoy starts up when the current Envoy has not yet fully
-  // initialized. The startup logic is quite complicated, and it's not worth trying to handle this
-  // in a finer way. This will cause the startup to fail with an error code early, without
-  // affecting any currently running processes. The process runner should try again later with some
-  // back off and with the same hot restart epoch number.
-  uint64_t old_flags = shmem->flags_.fetch_or(Flags::INITIALIZING);
-  if (old_flags & Flags::INITIALIZING) {
-    throw EnvoyException("previous envoy process is still initializing");
-  }
-
-  return *shmem;
-}
-
-void SharedMemory::initializeMutex(pthread_mutex_t& mutex) {
-  pthread_mutexattr_t attribute;
-  pthread_mutexattr_init(&attribute);
-  pthread_mutexattr_setpshared(&attribute, PTHREAD_PROCESS_SHARED);
-#ifdef PTHREAD_MUTEX_ROBUST
-  pthread_mutexattr_setrobust(&attribute, PTHREAD_MUTEX_ROBUST);
-#endif
-  pthread_mutex_init(&mutex, &attribute);
-}
-
-std::string SharedMemory::version() { return fmt::format("{}.{}", VERSION, sizeof(SharedMemory)); }
-
-#ifdef __linux__
 HotRestartImpl::HotRestartImpl(Options& options)
     : options_(options), shmem_(SharedMemory::initialize(options)), log_lock_(shmem_.log_lock_),
       access_log_lock_(shmem_.access_log_lock_), stat_lock_(shmem_.stat_lock_),
@@ -113,14 +43,6 @@ HotRestartImpl::HotRestartImpl(Options& options)
   RELEASE_ASSERT(rc != -1);
   UNREFERENCED_PARAMETER(rc);
 }
-
-#else // __linux__
-
-HotRestartImpl::HotRestartImpl(Options& options)
-    : shmem_(SharedMemory::initialize(options)), log_lock_(shmem_.log_lock_),
-      access_log_lock_(shmem_.access_log_lock_), stat_lock_(shmem_.stat_lock_) {}
-
-#endif // __linux__
 
 Stats::RawStatData* HotRestartImpl::alloc(const std::string& name) {
   // Try to find the existing slot in shared memory, otherwise allocate a new one.
@@ -148,8 +70,6 @@ void HotRestartImpl::free(Stats::RawStatData& data) {
 
   memset(&data, 0, sizeof(Stats::RawStatData));
 }
-
-#ifdef __linux__
 
 int HotRestartImpl::bindDomainSocket(uint64_t id) {
   // This actually creates the socket and binds it. We use the socket in datagram mode so we can
@@ -454,8 +374,6 @@ void HotRestartImpl::terminateParent() {
 void HotRestartImpl::shutdown() { socket_event_.reset(); }
 
 std::string HotRestartImpl::version() { return SharedMemory::version(); }
-
-#endif // __linux__
 
 } // namespace Server
 } // namespace Envoy
